requirements.txt
streamlit
pandas
matplotlib.pyplot
